[Music] Sync MT2 + perf content with upstream PR squashes
Adopt the source content from MatroskaTags2 (#28314) and PaPlayerPerformance
(#28323) after both PRs were squashed to single coherent commits today.
13 files updated to match the PR-squash state, with piers_qq private
features re-applied on top:

* FileDirectoryFactory: extend the PR's MUSIC::IsAudioBook()-based audiobook
  branch to also match .mp4 (piers_qq private), with the music-source guard
  applied to both .mkv and .mp4 dual-use containers.
* AudioBookFileDirectory / MusicInfoTagLoaderMatroska: keep
  SetBitsPerSample / SetCodec call sites uncommented (album codec/bps DB
  feature, piers_qq private).
* MusicCodecInfoFFmpeg: keep "dts" instead of "dca" for the default DTS
  profile (piers_qq private — master already renamed the codec name via
  PR #28140, but the new music DB column predates that rename in piers_qq).

Functional gains from the PR squashes that piers_qq now picks up:
* AudioDecoder: uint64_t cast on startThresholdBytes to avoid 32-bit
  wraparound on extreme hi-res multichannel (e.g. 32-bit/16ch/384 kHz).
* VideoPlayerCodec: planar-format bail-out + <8 bits guard in the
  m_frameSize fallback path (prevents division-by-zero in ReadPCM).
* AudioBookFileDirectory: minor unused-variable cleanup.
* clang-format applied to all final-state lines.

This keeps the piers_qq diff against future PR merges minimal, per the
piers_qq <-> master alignment workflow.

+++ xbmc/cores/paplayer/AudioDecoder.cpp
@@ -285,10 +285,14 @@ int CAudioDecoder::ReadSamples(int numsamples)
         // closer to that feel without sacrificing the resilience the remaining
         // buffer-fill provides during playback.
         constexpr unsigned int STARTUP_BUFFER_MS = 200;
-        const unsigned int startThresholdBytes =
-            (STARTUP_BUFFER_MS * (m_codec->m_bitsPerSample >> 3) *
-             m_codec->m_format.m_channelLayout.Count() *
-             m_codec->m_format.m_sampleRate) / 1000;
+        // 64-bit intermediate: 32-bit can wrap for extreme hi-res multichannel
+        // (e.g. 32-bit/16ch/384 kHz: 200 * 4 * 16 * 384000 ~= 4.92e9 > UINT_MAX).
+        const uint64_t startThresholdBytes =
+            (static_cast<uint64_t>(STARTUP_BUFFER_MS) *
+             static_cast<uint64_t>(m_codec->m_bitsPerSample >> 3) *
+             static_cast<uint64_t>(m_codec->m_format.m_channelLayout.Count()) *
+             static_cast<uint64_t>(m_codec->m_format.m_sampleRate)) /
+            1000;
 
         if (m_status == STATUS_QUEUING && m_pcmBuffer.getMaxReadSize() > startThresholdBytes)
         {
